pyproject.toml
[project]
name = "multimodal-inference-service"
version = "0.1.0"
description = "多模态推理服务：文生图与图片理解（FastAPI + WebSocket/HTTP）"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "fastapi>=0.115.0",
    "pillow>=12.1.1",
    "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/multimodal_inference_service"]

[tool.ruff]
target-version = "py310"
line-length = 100
force-exclude = true

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
